Split the setup out of the proxy service test
